feat(backend): add logs with first valued features (#13)
First feature with a real value.

Display logs with error / warning about the code so the end user can pick a log view which file it is and what is the issue.

Logged issues:
* component without propTypes
* components without displayName
* file not imported anywhere (Dead code)

Backend:
Add a second pass to use the data to enrich analytics:
We add logs which follow the idea to have { code, target } so we can display a message.
Apply rules on each file. One rule one code.

Frontend:
Add ViewFile to display file.
Refactor all other to be based on ViewFile.
Add logs in a SelectionList on top left (most important)
Force after a refresh to be redirected to the ViewRouter component. (best first screen)